Extract all common methods from dispatcher_haigha to dispatcher_common
The extracted methods are not tied to haigha and can be reused across multiple backends.
